iattendance/requirements.txt
backcall==0.1.0
bleach==2.1.4
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
colorlover==0.2.1
cufflinks==0.14.4
decorator==4.3.0
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.1.1
django-adaptors==0.2.5
django-excel==0.0.10
django-import-export==1.0.1
djangorestframework==3.8.2
entrypoints==0.2.3
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.7
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jdcal==1.4
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
lml==0.0.4
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.4.0
nbformat==4.4.0
nose==1.3.7
notebook==5.7.0
numpy==1.15.1
odfpy==1.3.6
openpyxl==2.5.6
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pd==0.0.1
pickleshare==0.7.4
plotly==3.2.1
prometheus-client==0.3.1
prompt-toolkit==1.0.15
pyexcel==0.5.9.1
pyexcel-io==0.5.9.1
pyexcel-webio==0.1.4
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.5
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
requests==2.19.1
retrying==1.3.3
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.11
tablib==0.12.1
terminado==0.8.1
testpath==0.3.1
texttable==1.4.0
tornado==5.1.1
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
xlrd==1.1.0
xlwt==1.3.0
